Improvements to Situation Gen and Physics Engine
Improved the physics simulator and added some more features like the
distance traveled before hitting the brakes and if the driver can stop.
Also improved the situation generator by creating the velocity and
acceleration in ft/s and ft/s^2 as opposed to mph. Overloaded the
constructor for both SimPhysics and the situation generator. Also added
comments for extra readability and clarity in units.