Formatting Improvements and Unit Standardization
Used String.format() statements in order to make the data easier to
read in the future. Also standardized all the units so that all
calculations are done in ft/s or ft/s^2. Also hand checked the
calculations so all calculations made from the code are 100% accurate.
Made some minor changes to the SimPhysics class which were mainly for
the sake of formatting.