Bug Fixes and Improvements to Situation Gen
After some testing, I found out that the equation used in
SimPhysics.stoppingTime(double vi, double a) was wrong, so I fixed
that. After looking at the data, I also saw that it would make sense to
have two different distance from light methods for the 50mph and 30mph
road since the stopping distance for the 50mph wasn’t nearly enough and
didn’t make sense logically. Also made some general improvements to the
tester so that it would be easier to collect data (instead of 10
samples, it now returns 20).