runYellow() initial commit
Added runYellow() to SimPhysics. The method determines if the car can
continue moving at constant acceleration and still make it past the
signal to the other side of the road. The code assumes that the length
of the intersection is 100 ft (average, or so I’m told). In order to
keep everything consistent, I added this data to the data table that is
printed when you run SimulatorTester. I don’t think that there are any
problems with the code right now since I hand checked the calculations
done by runYellow(). If you find an error, create a new issue and bring
it my attention.